+++ test/wallet-auction-test.js
@@ -848,7 +848,6 @@ describe('Wallet Auction', function() {
           ...historyOptions
         }));
 
-        console.log(txs.length);
         const wtxs = await wallet.toDetails(txs);
         for (const wtx of wtxs) {
           for (const output of wtx.outputs)